+++ recommender/Recommender.py
@@ -53,7 +53,7 @@ class Recommender:
             recorded_data (pd.DataFrame): The recorded metrics data for the current time window to simulate
         Returns:
             latest_time (datetime): The latest time of the performance data.
-            new_limit (int): The new number of cores to scale to.
+            new_limit (float): The new number of cores to scale to.
         """
         # # Do something here with the recorded_data to determine the new limit.
         # # The recorded_data is a pandas DataFrame with columns like timestamp, cpu_usage, etc.
